# DROOLS-6193 : Subnetwork Optimization - DO NOT MERGE YET