--- a/drools-core/src/main/java/org/drools/core/common/BaseNode.java
+++ b/drools-core/src/main/java/org/drools/core/common/BaseNode.java
@@ -128,7 +128,9 @@ public abstract class BaseNode
      * Attaches the node into the network. Usually to the parent <code>ObjectSource</code> or <code>TupleSource</code>
      */
     public void attach(BuildContext context) {
+        // do common shared code here, so it executes for all nodes
         doAttach(context);
+        // do common shared code here, so it executes for all nodes
     }
 
     public void doAttach(BuildContext context) {