… drools-core/src/main/java/org/drools/core/phreak/PhreakSubnetworkNotExistsNode.java
… drools-core/src/main/java/org/drools/core/phreak/SegmentPropagator.java
… drools-core/src/main/java/org/drools/core/reteoo/BaseLeftTuple.java
… drools-core/src/main/java/org/drools/core/reteoo/BetaNode.java
… drools-core/src/main/java/org/drools/core/reteoo/LeftInputAdapterNode.java
… drools-core/src/main/java/org/drools/core/reteoo/builder/CollectBuilder.java
… drools-core/src/main/java/org/drools/core/reteoo/builder/GroupElementBuilder.java
… drools-core/src/main/java/org/drools/core/reteoo/builder/NodeFactory.java
… drools-core/src/main/java/org/drools/core/reteoo/builder/PhreakNodeFactory.java
… drools-core/src/main/java/org/drools/core/spi/Tuple.java
… drools-core/src/main/java/org/drools/core/util/FastIterator.java
… drools-model/drools-model-compiler/src/test/java/org/drools/modelcompiler/AccumulateTest.java
… drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/ActivateAndDeleteOnListenerTest.java
… drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/BackwardChainingTest.java
… drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/SubnetworkTest.java
… drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/integrationtests/phreak/SegmentPropagationTest.java