All tests pass. Although it should add no changes in behaviour, I added some additional coverage in areas I was concerned about. I need to review changes, for cleanups, remove commented code etc. We also need to benchmark this, to make sure there are no actual regressions. It is still searching for the LT, we could optimise this further to record it in the constructor at the start - it won't matter much for small subnetworks, but would do if there was a bigger subnetwork.

When benchmark not only put lots of data inside of the subnetwork, make a lot of LTs prior to it, which we want to trigger updates on - which is also common in OP, with nesting of groupby's. These updates are the ones that typically cost the most, due to double update either side. Updates inside should be less problematic, if there is no chaining of groupbys. Need to benchmark not/exists/foreach and accumulate.