@mariofusco 2021-05-25
I ran [this benchmark](https://github.com/kiegroup/kie-benchmarks/pull/117) with and without this commit and got the following results

WITHOUT subnetwork optimization 
```
Benchmark                      (joinCount)  (nestedAccumulates)  (objectsPerSegment)  (perSegmentUpdate)  (segmentCount)  Mode  Cnt    Score   Error  Units
SubNetworkBenchmark.benchmark            2                 true                    5                true               3    ss  200    0.941 ± 0.067  ms/op
SubNetworkBenchmark.benchmark            2                 true                    5               false               3    ss  200    0.954 ± 0.054  ms/op
SubNetworkBenchmark.benchmark            2                 true                   15                true               3    ss  200    1.539 ± 0.165  ms/op
SubNetworkBenchmark.benchmark            2                 true                   15               false               3    ss  200    1.561 ± 0.089  ms/op
SubNetworkBenchmark.benchmark            2                false                    5                true               3    ss  200    1.830 ± 0.101  ms/op
SubNetworkBenchmark.benchmark            2                false                    5               false               3    ss  200    1.697 ± 0.091  ms/op
SubNetworkBenchmark.benchmark            2                false                   15                true               3    ss  200  137.876 ± 2.528  ms/op
SubNetworkBenchmark.benchmark            2                false                   15               false               3    ss  200   82.678 ± 1.578  ms/op
```

WITH subnetwork optimization 
```
Benchmark                      (joinCount)  (nestedAccumulates)  (objectsPerSegment)  (perSegmentUpdate)  (segmentCount)  Mode  Cnt   Score   Error  Units
SubNetworkBenchmark.benchmark            2                 true                    5                true               3    ss  200   1.050 ± 0.092  ms/op
SubNetworkBenchmark.benchmark            2                 true                    5               false               3    ss  200   0.994 ± 0.099  ms/op
SubNetworkBenchmark.benchmark            2                 true                   15                true               3    ss  200   2.147 ± 0.145  ms/op
SubNetworkBenchmark.benchmark            2                 true                   15               false               3    ss  200   2.043 ± 0.108  ms/op
SubNetworkBenchmark.benchmark            2                false                    5                true               3    ss  200   2.446 ± 0.108  ms/op
SubNetworkBenchmark.benchmark            2                false                    5               false               3    ss  200   1.973 ± 0.121  ms/op
SubNetworkBenchmark.benchmark            2                false                   15                true               3    ss  200  31.513 ± 0.952  ms/op
SubNetworkBenchmark.benchmark            2                false                   15               false               3    ss  200  22.349 ± 0.418  ms/op
```

These results have been obtained after one single fork with 100 warmup iterations and 200 benchmark iterations, so probably this would deserve a longer run before drawing any definitive conclusion. From this preliminary results, it is evident that this optimization has a huge positive impact on the longest running scenarios (sequential accumulates, 15 objects inserted per segment), while it seems to have some slightly negative consequences on the other shorter use cases.